Adds the async syntax to tagless (#508)
