+++ modules/async/async/shared/src/test/scala/free/AsyncTests.scala
@@ -24,7 +24,6 @@ import freestyle.free._
 import freestyle.free.implicits._
 import freestyle.free.async._
 import freestyle.free.async.implicits._
-import freestyle.async.implicits._
 
 import scala.concurrent.{ExecutionContext, Future}
